Rename SkillSearchBox to search box, it has no skill specific code and can be used elsewhere
